Move definition of GPUTypeChoices to avoid circular dependency
