+++ app/grandchallenge/algorithms/models.py
@@ -37,10 +37,10 @@ from grandchallenge.components.models import (  # noqa: F401
     ComponentInterfaceValue,
     ComponentJob,
     ComponentJobManager,
-    GPUTypeChoices,
     ImportStatusChoices,
     Tarball,
 )
+from grandchallenge.components.utils import GPUTypeChoices
 from grandchallenge.core.guardian import get_objects_for_group
 from grandchallenge.core.models import RequestBase, UUIDModel
 from grandchallenge.core.storage import (
