+++ app/grandchallenge/evaluation/forms.py
@@ -29,10 +29,10 @@ from grandchallenge.challenges.models import Challenge, ChallengeRequest
 from grandchallenge.components.forms import ContainerImageForm
 from grandchallenge.components.models import (
     ComponentInterface,
-    GPUTypeChoices,
     ImportStatusChoices,
 )
 from grandchallenge.components.tasks import assign_tarball_from_upload
+from grandchallenge.components.utils import GPUTypeChoices
 from grandchallenge.core.forms import (
     SaveFormInitMixin,
     WorkstationUserFilterMixin,
