+++ app/tests/algorithms_tests/factories.py
@@ -8,7 +8,7 @@ from grandchallenge.algorithms.models import (
     AlgorithmUserCredit,
     Job,
 )
-from grandchallenge.components.models import GPUTypeChoices
+from grandchallenge.components.utils import GPUTypeChoices
 from tests.components_tests.factories import ComponentInterfaceValueFactory
 from tests.factories import (
     ImageFactory,
