+++ app/tests/algorithms_tests/test_models.py
@@ -19,8 +19,8 @@ from grandchallenge.components.models import (
     CIVData,
     ComponentInterface,
     ComponentInterfaceValue,
-    GPUTypeChoices,
 )
+from grandchallenge.components.utils import GPUTypeChoices
 from tests.algorithms_tests.factories import (
     AlgorithmFactory,
     AlgorithmImageFactory,
