+++ app/tests/algorithms_tests/test_tasks.py
@@ -17,12 +17,12 @@ from grandchallenge.algorithms.tasks import (
 from grandchallenge.components.models import (
     ComponentInterface,
     ComponentInterfaceValue,
-    GPUTypeChoices,
 )
 from grandchallenge.components.tasks import (
     add_image_to_component_interface_value,
     validate_docker_image,
 )
+from grandchallenge.components.utils import GPUTypeChoices
 from grandchallenge.notifications.models import Notification
 from tests.algorithms_tests.factories import (
     AlgorithmFactory,
