+++ app/tests/components_tests/test_amazon_sagemaker_training_backend.py
@@ -18,7 +18,7 @@ from grandchallenge.components.backends.exceptions import (
     TaskCancelled,
 )
 from grandchallenge.components.backends.utils import LOGLINES
-from grandchallenge.components.models import GPUTypeChoices
+from grandchallenge.components.utils import GPUTypeChoices
 from grandchallenge.evaluation.models import Evaluation, Method
