+++ app/tests/evaluation_tests/factories.py
@@ -1,6 +1,6 @@
 import factory
 
-from grandchallenge.components.models import GPUTypeChoices
+from grandchallenge.components.utils import GPUTypeChoices
 from grandchallenge.evaluation.models import (
     CombinedLeaderboard,
     Evaluation,
